+++ Changelog.md
@@ -16,7 +16,8 @@ Bugfixes:
  * Type checker: fixed crash related to invalid literal numbers.
  * Type Checker: ``super.x`` does not look up ``x`` in the current contract.
  * Code generator: expect zero stack increase after ``super`` as an expression.
- * Code Generator: fixed an internal compiler error for ``L.Foo`` for ``enum Foo`` defined in library ``L``.
+ * Code Generator: fix an internal compiler error for ``L.Foo`` for ``enum Foo`` defined in library ``L``.
+ * Code generator: allow inheritance of ``enum`` definitions.
  * Inline assembly: support the ``address`` opcode.
  * Inline assembly: fix parsing of assignment after a label.
  * Inline assembly: external variables of unsupported type (such as ``this``, ``super``, etc.)